// src/app/slices/city/types.ts
export interface Location {
  name: string;
  region: string;
  country: string;
  lat: number;
  lon: number;
  tz_id: string;
  localtime_epoch: number;
  localtime: string;
}

export interface Condition {
  text: string;
  icon: string;
  code: number;
}

export interface Current {
  last_updated_epoch: number;
  last_updated: string;
  temp_c: number;
  temp_f: number;
  is_day: number;
  condition: Condition;
  wind_mph: number;
  wind_kph: number;
  wind_degree: number;
  wind_dir: string;
  pressure_mb: number;
  pressure_in: number;
  precip_mm: number;
  precip_in: number;
  humidity: number;
  cloud: number;
  feelslike_c: number;
  feelslike_f: number;
  vis_km: number;
  vis_miles: number;
  uv: number;
  gust_mph: number;
  gust_kph: number;
}
// ...
export interface Forecastday {
  date: string;
  date_epoch: number;
  day: Day;
  astro: Astro;
  hour: Hour[];
}

export interface Forecast {
  forecastday: Forecastday[];
}

export interface WeatherCityResponse {
  location: Location;
  current: Current;
  forecast: Forecast;
}

export interface SportEvent {
  stadium: string;
  country: string;
  region: string;
  tournament: string;
  start: string;
  match: string;
}

export interface EventsResponse {
  football: SportEvent[];
  cricket: SportEvent[];
  golf: SportEvent[];
}

export interface SportEventData extends SportEvent {
  sport: string;
}

export interface AutoComplete {
  id: number;
  name: string;
  region: string;
  country: string;
  lat: number;
  lon: number;
  url: string;
}

export interface CityState {
  weather: WeatherCityResponse | null;
  weatherHistory: WeatherCityResponse | null;
  favoriteCities: WeatherCityResponse[];
  favoriteEvents: SportEventData[];
  sports: SportEventData[];
  autocomplete: AutoComplete[];
  cityStatus: itemStatus;
  autocompleteStatus: itemStatus;
  sportsStatus: itemStatus;
  weatherHistoryStatus: itemStatus;
  historyError: string | null;
  searchError: string | null;
  sportsError: string | null;
  weatherError: string | null;
}